Use 64 byte long salt
The 2nd line in the state file is the salt now, and the 3rd is the
hashed response.

Users are generating a random salt when they're setting up their state
files. We only use the salt when reading the state file to compute the
hash and compare them.

Bumped STATE_FILE_DATA_SIZE accordingly.

Updated README and manual accordingly.